+++ projects/ngx-mat-lib/src/lib/xmat-models/xmat-http-config.model.ts
@@ -1,11 +1,11 @@
 import {HttpParams} from "@angular/common/http";
-import {XmatDelayedHttpParams} from "./xmat-delayed-http-params.model";
 import {XmatRestVerbs} from "./xmat-rest-verbs.model";
+import {XmatHttpParams} from "./xmat-http-params.model";
 
 export interface XmatHttpConfig {
     method: XmatRestVerbs;
     url: string;
     data?: object;
     queryable: boolean;
-    params?: HttpParams | XmatDelayedHttpParams | any;
+    params?: HttpParams | XmatHttpParams | any;
 }