+++ projects/ngx-mat-lib/src/lib/xmat-services/xmat-rest.service.ts
@@ -1,7 +1,7 @@
 import {Injectable} from "@angular/core";
 import {HttpClient, HttpParams} from "@angular/common/http";
-import {Observable} from "rxjs";
-import {forkJoin} from "rxjs";
+import {Observable, forkJoin, of} from "rxjs";
+import {map} from "rxjs/operators";
 import {XmatConstantsService} from "./xmat-constants.service";
 import {
     XmatFile,
@@ -9,8 +9,8 @@ import {
     XmatResponseTypes,
     XmatRestVerbs
 } from "../xmat-models/index";
-import * as _ from "lodash";
 import {XmatGenericObject} from "../xmat-models/index";
+import * as _ from "lodash";
 
 /**
  * XMAT REST BY TONY SAMPERI
@@ -95,19 +95,30 @@ export class XmatRestService {
     }
 
     $all(configs: XmatHttpConfig[]): Observable<any[]> {
-        const queue = [];
+        const queue: Observable<any>[] = [];
         _.each(configs, (c: XmatHttpConfig) => {
             queue.push(this.$http(c));
         });
+
         return forkJoin(queue);
     }
 
-    $all(configs: XmatGenericObject<XmatHttpConfig>): Observable<XmatGenericObject<XmatHttpConfig>> {
-        const queue = [];
-        _.each(configs, (c: XmatHttpConfig) => {
+    $allMap(configs: XmatGenericObject<XmatHttpConfig>): Observable<any> {
+        const queue: Observable<any>[] = [];
+        const queueKeys: Array<number | string> = [];
+        _.each(configs, (c: XmatHttpConfig, key: string | number) => {
             queue.push(this.$http(c));
+            queueKeys.push(key);
         });
-        return forkJoin(queue);
+        return forkJoin(queue)
+        .pipe(map((raw: any[]) => {
+            const mapped = {};
+            _.each(raw, (value, index) => {
+                mapped[queueKeys[index]] = value;
+            });
+
+            return mapped;
+        }));
     }
 
     $http<T>(config: XmatHttpConfig = this._generateHttpConfig()): Observable<T> {